Move GetMimeType extension to Extensions folder and namespace
